Revert "added class ListIterator, returned from_array (#3)"
This reverts commit f80a65dca55174c850171dfa8a62b341cc42edef.
