--- progress.ini
@@ -1,14 +1,0 @@
-not sure:
-__getitem__ using iterator
-
-added:
-self.length
-__len__, __iter__, __next__, __getitem__, __setitem__, __eq__
-index_transform, negative indices
-insert
-changed find_node
-
-to do list:
-DLL
-stack
-public, private
